requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
boto3==1.16.61
botocore==1.19.61
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-appconf==1.0.4
django-ckeditor==6.0.0
django-filter==2.4.0
django-heroku==0.3.1
django-image-cropping==1.5.0
django-jquery==3.1.0
django-js-asset==1.2.2
django-simple-metatags==2.0.1
django-sitemap-generate==0.2.1
django-storages==1.11.1
easy-thumbnails==2.7.1
gunicorn==20.0.4
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
